[refactor] BaseTableViewCell 추가
- 기존 tableView cell 상속 받는 구조로 수정
