+++ src/smooth-ios/smooth-ios/Modules/Friend/SubView/FriendListViewCell.swift
@@ -10,10 +10,58 @@ import Then
 import SnapKit
 import RxSwift
 
-class FriendListViewCell: UITableViewCell {
-    var disposeBag = DisposeBag()
+class FriendListViewCell: BaseTableViewCell {
+    static let identifier = "\(FriendListViewCell.self)"
     
-    static let identifier = "FriendCell"
+    let profileImg = UIImageView().then {
+        $0.layer.cornerRadius = 15
+        $0.layer.masksToBounds = true
+        $0.clipsToBounds = true
+    }
+    
+    let nameLabel = UILabel().then {
+        $0.text = "두리짱"
+        $0.textColor = .white
+        $0.translatesAutoresizingMaskIntoConstraints = false
+    }
+    
+    let stateLabel = UILabel().then {
+        $0.textColor = .white
+        $0.font = UIFont.systemFont(ofSize: 9, weight: .thin)
+        $0.translatesAutoresizingMaskIntoConstraints = false
+    }
+    
+    let acceptButton = UIButton().then {
+        $0.backgroundColor = .backgroundDarkGray
+        $0.setImage(UIImage(systemName: "checkmark")?.withTintColor(.online!, renderingMode: .alwaysOriginal), for: .normal)
+        
+        $0.layer.cornerRadius = 15
+        $0.layer.masksToBounds = true
+    }
+    
+    let rejectButton = UIButton().then {
+        $0.backgroundColor = .backgroundDarkGray
+        $0.setImage(UIImage(systemName: "multiply")?.withTintColor(.red!, renderingMode: .alwaysOriginal), for: .normal)
+        
+        $0.layer.cornerRadius = 15
+        $0.layer.masksToBounds = true
+    }
+    
+    let messageButton = UIButton().then {
+        $0.backgroundColor = .backgroundDarkGray
+        $0.setImage(UIImage(systemName: "bubble.left.fill")?.withTintColor(.greyple!, renderingMode: .alwaysOriginal), for: .normal)
+        
+        $0.layer.cornerRadius = 15
+        $0.layer.masksToBounds = true
+    }
+    
+    let callingButton = UIButton().then {
+        $0.backgroundColor = .backgroundDarkGray
+        $0.setImage(UIImage(systemName: "phone.fill")?.withTintColor(.greyple!, renderingMode: .alwaysOriginal), for: .normal)
+        
+        $0.layer.cornerRadius = 15
+        $0.layer.masksToBounds = true
+    }
     
     override func prepareForReuse() {
         super.prepareForReuse()
@@ -29,66 +77,6 @@ class FriendListViewCell: UITableViewCell {
         
         self.profileImg.subviews.forEach {$0.removeFromSuperview()}
         self.disposeBag = DisposeBag()
-    }
-    
-    let profileImg = UIImageView().then {
-        $0.layer.cornerRadius = 15
-        $0.layer.masksToBounds = true
-        $0.clipsToBounds = true
-    }
-    
-    let nameLabel = UILabel().then {
-        $0.text = "두리짱"
-        $0.textColor = .white
-        $0.translatesAutoresizingMaskIntoConstraints = false
-    }
-
-    let stateLabel = UILabel().then {
-        $0.textColor = .white
-        $0.font = UIFont.systemFont(ofSize: 9, weight: .thin)
-        $0.translatesAutoresizingMaskIntoConstraints = false
-    }
-    
-    let acceptButton = UIButton().then {
-        $0.backgroundColor = .backgroundDarkGray
-        $0.setImage(UIImage(systemName: "checkmark")?.withTintColor(.online!, renderingMode: .alwaysOriginal), for: .normal)
-    
-        $0.layer.cornerRadius = 15
-        $0.layer.masksToBounds = true
-    }
-    
-    let rejectButton = UIButton().then {
-        $0.backgroundColor = .backgroundDarkGray
-        $0.setImage(UIImage(systemName: "multiply")?.withTintColor(.red!, renderingMode: .alwaysOriginal), for: .normal)
-    
-        $0.layer.cornerRadius = 15
-        $0.layer.masksToBounds = true
-    }
-    
-    let messageButton = UIButton().then {
-        $0.backgroundColor = .backgroundDarkGray
-        $0.setImage(UIImage(systemName: "bubble.left.fill")?.withTintColor(.greyple!, renderingMode: .alwaysOriginal), for: .normal)
-    
-        $0.layer.cornerRadius = 15
-        $0.layer.masksToBounds = true
-    }
-    
-    let callingButton = UIButton().then {
-        $0.backgroundColor = .backgroundDarkGray
-        $0.setImage(UIImage(systemName: "phone.fill")?.withTintColor(.greyple!, renderingMode: .alwaysOriginal), for: .normal)
-
-        $0.layer.cornerRadius = 15
-        $0.layer.masksToBounds = true
-    }
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        layout()
-        
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
     }
     
     func bind(friend: Friend) {
@@ -123,15 +111,15 @@ class FriendListViewCell: UITableViewCell {
             self.callingButton.isHidden = true
             
             
-        // TODO: 친구 상태값(온라인, 오프라인, 자리비움 등등) 노출
+            // TODO: 친구 상태값(온라인, 오프라인, 자리비움 등등) 노출
         case .accept:
             self.rejectButton.isHidden = true
             self.acceptButton.isHidden = true
-
+            
         case .ban:
             self.rejectButton.isHidden = true
             self.acceptButton.isHidden = true
-
+            
         case .none:
             self.acceptButton.isHidden = true
             self.messageButton.isHidden = true
@@ -149,7 +137,7 @@ class FriendListViewCell: UITableViewCell {
         // wait(친구 요청 응답 기다리는 중)
     }
     
-    func layout() {
+    override func setup() {
         [
             profileImg, nameLabel, stateLabel,
             rejectButton, acceptButton,
@@ -157,7 +145,9 @@ class FriendListViewCell: UITableViewCell {
         ].forEach {
             self.contentView.addSubview($0)
         }
-        
+    }
+    
+    override func bindConstraints() {
         profileImg.snp.makeConstraints {
             $0.centerY.equalToSuperview()
             $0.left.equalToSuperview().offset(20)
